Moved section inside page component
