+++ src/pages/Body.js
@@ -5,9 +5,10 @@ import styled from 'styled-components'
 import { endpoints, colors } from '../constants'
 
 import PageHeader from '../components/PageHeader'
-import CelestialObjectSection from '../components/CelestialObjectSection'
 import CelestialObjectCardsContainer from '../components/CelestialObjectCardsContainer'
 import CelestialObjectCard from '../components/CelestialObjectCard'
+import CelestialObject from '../components/CelestialObject'
+import CelestialObjectInformation from '../components/CelestialObjectInformation'
 
 const BodyPage = () => {
   const [ isLoading, setIsLoading ] = useState(false)
@@ -61,7 +62,12 @@ const BodyPage = () => {
   return (
     <>
       <PageHeader title={data.englishName || 'Unknown'} />
-      <CelestialObjectSection data={data} color={colors?.[data.id] || colors.moon} />
+      
+      <StyledCelestialObjectSection>
+        <StyledCelestialObject color={colors?.[data.id] || colors.moon} />
+        <StyledCelestialObjectInformation data={data} />
+      </StyledCelestialObjectSection>
+
       <StyledMoonsSection>
         <h2>Moons</h2>
         {data.moons && <CelestialObjectCardsContainer>
@@ -74,6 +80,30 @@ const BodyPage = () => {
     </>
   )
 }
+
+const StyledCelestialObjectSection = styled.section`
+  display: flex;
+  flex-direction: column;
+  gap: 2em;
+  padding-block: 2em;
+
+  @media (min-width: 48em) {
+    flex-direction: row;
+    justify-content: space-between;
+  }
+`
+
+const StyledCelestialObject = styled(CelestialObject)`
+  @media (min-width: 48em) {
+    max-width: 40%;
+  }
+`
+
+const StyledCelestialObjectInformation = styled(CelestialObjectInformation)`
+  @media (min-width: 48em) {
+    min-width: 50%;
+  }
+`
 
 const StyledMoonsSection = styled.section`
   padding-block: 2em;
